Version inconsistencies fixed. Add formal lang
There were two different version strings in the code. Increased the version number in train of fixing this.

Also depending on Moodle settings the first name is in field element 0, as already mentioned by TRMSC. Fixed this for our Moodle (version 4.0.4) and it's settings. Also included a way to have German formal language, with first and last name.